requirement.txt
Flask==2.0.2
Flask-Cors==3.0.10
python-dotenv==0.19.0
termcolor==1.1.0
groq